Modify NSErrorHelper to common space for Auth use
