+++ Assets/AppCenter/Plugins/iOS/Core/Utility/NSErrorHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f73030502bdd74d458f10a7c1a597e3b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
