+++ Assets/AppCenter/Plugins/iOS/Core/Utility/NSErrorHelper.h
