+++ Assets/AppCenter/Plugins/iOS/Core/Utility/NSErrorHelper.h.meta
@@ -1,0 +1,24 @@
+fileFormatVersion: 2
+guid: 23252b01a9d2f4f40ba6f2f0db9efbd5
+PluginImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  iconMap: {}
+  executionOrder: {}
+  isPreloaded: 0
+  isOverridable: 0
+  platformData:
+  - first:
+      Any: 
+    second:
+      enabled: 1
+      settings: {}
+  - first:
+      Editor: Editor
+    second:
+      enabled: 0
+      settings:
+        DefaultValueInitialized: true
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
