+++ Assets/AppCenter/Plugins/iOS/Core/Utility/NSErrorHelper.mm
