+++ Assets/AppCenter/Plugins/iOS/Core/Utility/NSErrorHelper.mm.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d3887caf16d5f413eb539ba94423b969
+guid: b122ae60c2ba54e659d9568ad78400e7
 PluginImporter:
   externalObjects: {}
   serializedVersion: 2
