+++ Assets/AppCenter/Plugins/iOS/Crashes/Utility/NSErrorHelper.cs
@@ -3,12 +3,11 @@
 
 #if UNITY_IOS && !UNITY_EDITOR
 using System;
-using System.Runtime.InteropServices;
 using Exception = Microsoft.AppCenter.Unity.Crashes.Models.Exception;
 
-namespace Microsoft.AppCenter.Unity.Crashes.Utility
+namespace Microsoft.AppCenter.Unity.Internal.Utility
 {
-    public static class NSErrorHelper
+    public static partial class NSErrorHelper
     {
         public static Exception Convert(IntPtr nsErrorPtr)
         {
@@ -21,15 +20,6 @@ namespace Microsoft.AppCenter.Unity.Crashes.Utility
             var description = app_center_unity_nserror_description(nsErrorPtr);
             return new Exception(string.Format("Domain: {0}, error code: {1}, description: {2}", domain, errorCode, description), string.Empty);
         }
-
-        [DllImport("__Internal")]
-        private static extern string app_center_unity_nserror_domain(IntPtr error);
-
-        [DllImport("__Internal")]
-        private static extern long app_center_unity_nserror_code(IntPtr error);
-
-        [DllImport("__Internal")]
-        private static extern string app_center_unity_nserror_description(IntPtr error);
     }
 }
 #endif
